+++ Unison-Demo/UnisonDemo/ContentView.swift
@@ -40,18 +40,6 @@ extension ContentView {
     }
 }
 
-struct LazyView<Content: View>: View {
-    let build: () -> Content
-    
-    init(_ build: @autoclosure @escaping () -> Content) {
-        self.build = build
-    }
-    
-    var body: Content {
-        build()
-    }
-}
-
 #if DEBUG
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
