+++ Unison-Demo/UnisonDemo/Helper/LazyView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct LazyView<Content: View>: View {
+    let build: () -> Content
+    
+    init(_ build: @autoclosure @escaping () -> Content) {
+        self.build = build
+    }
+    
+    var body: Content {
+        build()
+    }
+}
